requirements.txt
charset-normalizer==3.4.4
lxml==6.0.2
pillow==12.0.0
python-docx==1.2.0
reportlab==4.4.5
setuptools==80.9.0
typing_extensions==4.15.0
